+++ src/torchoutil/nn/modules/__init__.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+
 from .activation import *
 from .crop import *
 from .mask import *
@@ -8,6 +9,6 @@ from .multiclass import *
 from .multilabel import *
 from .numpy import *
 from .pad import *
-from .repeat import *
 from .tensor import *
+from .transform import *
 from .typed import *