+++ src/torchoutil/nn/modules/transform.py
@@ -1,9 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+from typing import Iterable, Union
+
 from torch import Tensor, nn
 
-from torchoutil.nn.functional.repeat import repeat_interleave_nd
+from torchoutil.nn.functional.transform import repeat_interleave_nd, resample_nearest
 from torchoutil.utils.collections import dump_dict
 
 
@@ -43,3 +45,20 @@ class RepeatInterleaveNd(nn.Module):
 
     def extra_repr(self) -> str:
         return dump_dict(dict(repeats=self.repeats, dim=self.dim))
+
+
+class ResampleNearest(nn.Module):
+    def __init__(
+        self,
+        rates: Union[float, Iterable[float]],
+        dims: Union[int, Iterable[int]] = -1,
+    ) -> None:
+        super().__init__()
+        self.rates = rates
+        self.dims = dims
+
+    def forward(self, x: Tensor) -> Tensor:
+        return resample_nearest(x, self.rates, self.dims)
+
+    def extra_repr(self) -> str:
+        return dump_dict(dict(rates=self.rates, dims=self.dims))