+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Option `padding_idx` to `indices_to_onehot` function.
 - Function `dict_list_to_list_dict` to collections utils.
 - Class `PickleDataset` and function `pack_to_pickle` to utils.
+- Class `ResampleNearest` and function `resample_nearest`.
 
 ### Fixed
 - Remove invalid test file.